Move makeSchedule tests from core/task_medium_test.go to core/schedule_small_test.go
